Add flowchart useMaxWidth config parameter
The useMaxWidth parameter controls whether the diagram will be scaled to
fit the available width. The default for this parameter is true, meaning
that the diagram will be scaled. For very wide diagrams, this can result
in an unreadably small diagram. Setting useMaxWidth to false allows the
diagram to remain at a fixed sized. Scrollbars can be added to the area,
if desired, with:

.ext-mermaid > div {
	overflow: scroll;
}